products/discord_bot/requirements.txt
discord.py==2.3.1
pandas==2.0.3
python-dotenv==1.0.0
